requirements/base.txt
aiofiles==0.7.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fake-useragent==0.1.11
fastapi==0.65.1
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.2.3
python-decouple==3.4
requests==2.25.1
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
